requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
idna==3.3
peewee==3.15.0
peewee-migrations==0.3.32
python-dotenv==0.20.0
requests==2.28.1
urllib3==1.26.10
